project microservice
project portion of the group project, deals with handling user, project, and project to resource related services.  Can register/login users.  Add/delete/update a project, and add a formula/extra column to the project which is owned by the user.  Microservice connects to the product microservice through eureka to allow adding a resource to a specific project